requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.13
psycopg2-binary>=2.8,<3.0